microservicio-cuentas/locust/requirements.txt
locust==2.31.8
mysql-connector-python==9.1.0
